refactor, removed unsed code
