+++ frontend/src/components/index.ts
@@ -1,1 +1,3 @@
 export * from './Navbar'
+export * from './CodeHighlight'
+export * from './PromptingTemplate'
